Set error fix, temp over request fix, LDR enable integration
There was an error where a reserved word used for variable name and this caused halt on SET menu exit.
Temperature conversion was checked on every loop and this caused temp rise on sensor chip. added an extra control and variable to set check interval.
Implemented LDR enable. this way users change of brightness on settings will persist and no pin enable and reads are done on them